Disable container log by default
NSO process running in foreground in verbose mode will output the log
messages to stdout, which is picked up by the container runtime. To
avoid overflowing the storage when the container runtime (Docker) is
misconfigured, let's disable this by default. Logs are persisted and
rotated in the `/log` volume anyway.